requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asyncio==3.4.3
certifi==2025.4.26
click==8.1.8
confluent-kafka==2.10.0
distro==1.9.0
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jiter==0.9.0
linecache2==1.0.0
openai==1.76.0
packaging==25.0
pluggy==1.5.0
prometheus_client==0.21.1
# protobuf version is intentionally unpinned:
# opentelemetry-proto 1.27.0 requires protobuf<5.0,>=3.19
# temporalio 1.11.0 requires protobuf>=3.20
# pip will resolve to protobuf ~4.x which satisfies both
protobuf>=3.20,<5.0
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.1.0
PyYAML==6.0.2
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
temporalio==1.11.0
tqdm==4.67.1
traceback2==1.4.0
types-protobuf>=4.0,<5.0
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
motor==3.7.0

# Optional agent integrations (install as needed)
# langchain-core>=0.3.0
# langchain-openai>=0.2.0
# llama-index-core>=0.11.0
# llama-index-llms-openai>=0.3.0

# Observability
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-exporter-otlp-proto-grpc==1.27.0

# Cognitive layer — agentic_codex (bundled in Agent-Components/)
# Local dev:  pip install -e ./Agent-Components
# Docker:     handled in Dockerfile after COPY . . (see Dockerfile.orchestrator)
